uncoupling functions to classes components
