+++ src/app/components/Input/Input.js
@@ -1,0 +1,20 @@
+import { css } from 'aphrodite'
+import styles from './../../styles';
+
+export class InputToDoItemComponent {
+    static renderInput () {
+        return `<div class="todo__input ${css(styles.divFullWidth, styles.divAlignFlex, styles.divAddTodo)}">
+            <button class="${css(styles.buttonAddTodo)}" id="addTodo">
+                <i class="add circle icon ${css(styles.iconAddTodoButton)}"></i>
+            </button>
+            <input placeholder="Add a Task" class="${css(styles.fullWidth, styles.inputAddTodo)}" type="text" id="todoInput">
+            <label class="${css(styles.fieldSelectSeverity)} severity">
+                <p class="${css(styles.textSeveritySelected)}">Set severity</p>
+                <select class="${css(styles.selectSeverity)}" id="set-severity">
+                    <option value="important">important</option>
+                    <option value="urgent">urgent</option>
+                </select>
+            </label>
+        </div>`;
+    }
+}
