+++ src/app/components/View.js
@@ -1,0 +1,5 @@
+export default class Component {
+    render (el, data) {
+        el.innerHTML = data;
+    }
+}
